cpu usage first implementation

+++ models/metricModel.go
@@ -6,7 +6,7 @@ type Metrics struct {
 	gorm.Model
 	ID                  uint    `gorm:"primaryKey;autoIncrement"`
 	TimeSpent           float64 //second
-	CpuUsage            uint
+	CpuUsage            float64
 	RamUsage            uint
 	RequestCount        uint
 	succeedRequestCount uint
